requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
colorama==0.4.6
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lark==1.3.1
openpyxl==3.1.5
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
python-hcl2==7.3.1
regex==2025.11.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
